ㅁ

+++ Assets/HongYeseong/script/ScripterScript/CharacterStats.cs
@@ -6,10 +6,10 @@ public class CharacterStats : ScriptableObject
     public string characterName;
 
     [Header("Stats")] public int health; // 체력
-    public int stamina; // 스테미너
-    public int power; // 힘
-    public int defense; // 방어력
-    public int intelligence; // 지식
+    public float stamina; // 스테미너
+    public float power; // 힘
+    public float defense; // 방어력
+    public float intelligence; // 지식
     public float speed; // 속도
     public float projectileSpeed;
     public float cooldown;
